requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
matplotlib>=3.7
pytest>=7.3
jpype1>=1.5.0
py4j>=0.10.9
sfctools>=0.6
